requirements.txt
altair==4.1.0
appdirs==1.4.4
argcomplete==1.10.0
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
backcall==0.2.0
base58==2.0.1
bleach==3.1.5
blinker==1.4
blis==0.4.1
boto3==1.15.11
botocore==1.18.11
bs4==0.0.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==5.0.0
crayons==0.4.0
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
entrypoints==0.3
enum-compat==0.0.3
et-xmlfile==1.0.1
extract-msg==0.23.1
filelock==3.0.12
future==0.18.2
futures==3.1.1
isort==5.5.1
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.2
numpy==1.19.1
olefile==0.46
openpyxl==3.0.5
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
patsy==0.5.1
pdfplumber==0.5.24
pendulum==2.1.2
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
plotly==4.12.0
pluggy==0.13.1
praw==7.1.0
prawcore==1.5.0
preshed==3.0.2
prettyprint==0.1.5
prettytable==2.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
py==1.9.0
pyarrow==1.0.1
pycparser==2.20
pycryptodome==3.9.9
pydeck==0.5.0b1
PyDictionary==2.0.1
pyecharts==1.9.0
Pygments==2.6.1
pylint==2.6.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyquery==1.4.1
pyrsistent==0.16.0
pytest==6.1.1
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.1
pytzdata==2020.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.7.14
requests==2.24.0
retry==0.9.2
retrying==1.3.3
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.3
seaborn==0.11.0
selenium==3.141.0
Send2Trash==1.5.0
sentencepiece==0.1.91
smmap==3.0.4
sortedcontainers==2.3.0
statsmodels==0.12.1
streamlit==0.71.0
streamlit-echarts==0.1.0
streamlit-metrics==0.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.4.1
threadpoolctl==2.1.0
tokenizers==0.9.3
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
transformers==3.5.1
tzlocal==1.5.1
urllib3==1.25.10
validators==0.18.1
virtualenv==20.0.31
Wand==0.6.3
wasabi==0.8.0
watchdog==0.10.3
wcwidth==0.2.5
webdriver-manager==3.2.2